requirements.txt
opencv-python>=4.8.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
numpy>=1.24.0
Pillow>=10.0.0
customtkinter>=5.2.0
matplotlib>=3.7.0
joblib>=1.3.0
seaborn>=0.12.0
